feat: add Routes and Route to the App component and create Main page
